refactor(cli): co-locate TestPlan resolver with its +ManifestMapper
Moves \`TestPlan.resolve(entries:...)\` and the \`appendPathEntry\` helper
out of \`TestAction+ManifestMapper\` and into the file that already
owns the TestPlan mapping (\`TestPlan+ManifestMapper\`), so both entry
points (\`TestPlan.from(path:...)\` and \`TestPlan.resolve(...)\`) live
together.
